refactor(visio): extraire jitsiRoom + helpers de salle hors du service (G8)
Sort la logique pure de construction de salle/lien Jitsi (generateRoomLink,
generateRoomId, buildRoomName, sanitizeForUrl) vers src/utils/jitsiRoom.js.
jitsiService délègue à ces helpers : API publique strictement identique
(mêmes signatures, mêmes URLs, mêmes logs). Effet de bord cleanupExpired
au chargement et itérations localStorage inchangés. Service 366 -> 298 lignes.
